rename functions to represent functions
